Display banner from Pluto.__init__, not during precompile
Display banner only when `$JULIA_DEPOT_PATH/pluto_notebooks/` is
created, and create that folder when the module initializes.

This preserves the fact that the banner is only displayed once, while
it now no longer displays it when the package is compiled.
Compile-time display became problematic as that is counted by `Pkg` as
a warning, and users might therefore think something went wrong.

Note that this now creates `$JULIA_DEPOT_PATH/pluto_notebooks/` much
earlier, whereas previously that folder was only created when the user
saved a notebook there.

fixes #548